fix: sales page components
